Fix text resizing bug.
According to the documentation, the `-webkit-text-size-adjust` CSS property only applies to Safari on iPhone. Because of this, we had assumed that setting it to none (as we do for resolutions below 768px wide) would only affect MobileSafari.

In fact, this property affects all versions of Safari, even desktop Safari, with the effect of disabling page zoom for narrow browser windows. Which is not what we want. If, on the other hand, we set the value of the property to 100%, we effectively stop Safari from resizing the text, but do not disable browser zoom.