Fix button alignment issue once and for all.
Just don't use Microsoft C-fonts on buttons: they sit too high compared to other fonts in the stack and and are hard to align, especially considering that some users may not have them and fall back to a font that sits lower. Use Verdana, with Tahoma as a backup, as the default button font instead.

Also remove the bit about generating the sassy-buttons CSS file from the Readme, since we'll be doing that ourselves every time we upgrade the Gem.